backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pydantic==2.9.2
pymongo==4.8.0
pinecone-client==4.1.0
openai==1.55.3
httpx==0.27.2
tenacity==9.0.0
pytest==8.3.3
